requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
# psycopg2
# Psycopg is the most popular PostgreSQL database adapter for the Python programming language.
drf-spectacular>=0.15.1,<0.16

